refactor: change file extension from .tsx to .ts for data file
